#29 fix: 商品画像の表示切り替えコンポーネントを追加した
