+++ src/components/ImageViewer/ImageViewer.module.scss
@@ -1,0 +1,31 @@
+.image__viewer {
+  display: flex;
+  gap: 8px;
+
+  &__main {
+    width: 400px;
+    height: 400px;
+    cursor: pointer;
+
+    &__image {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+  }
+
+  &__sub {
+    width: 120px;
+    height: 120px;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    cursor: pointer;
+
+    &__image {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+  }
+}
