+++ src/components/ImageViewer/ImageViewer.type.ts
@@ -1,0 +1,4 @@
+export type ImageType = {
+  url: string;
+  alt: string;
+};
