+++ src/components/ItemDetail/ItemDetail.module.scss
@@ -19,22 +19,6 @@
   display: flex;
   flex-direction: column;
   gap: 16px;
-
-  .mainImage {
-    width: 100%;
-    height: auto;
-  }
-
-  .subImages {
-    display: flex;
-    gap: 8px;
-
-    .subImage {
-      width: 80px;
-      height: 80px;
-      object-fit: cover;
-    }
-  }
 }
 
 .infoSection {
